chore: rename files and folders for `/users` RDB
