+++ entrypoints/content/firebase/realtime-database/muted/set.ts
@@ -1,0 +1,2 @@
+// muteUser
+// unmuteUser
