+++ entrypoints/content/utils/thoroughAuthCheck.ts
@@ -1,6 +1,6 @@
 // Packages:
 import returnable from 'utils/returnable'
-import { getRDBUser } from '../firebase/realtime-database/user/get'
+import { getRDBUser } from '../firebase/realtime-database/users/get'
 import thoroughUserDetailsCheck from 'utils/thoroughUserDetailsCheck'
 
 // Typescript:
